winre uninstall fix

+++ Rectify11Installer/Core/Rectify11Installer.cs
@@ -403,11 +403,14 @@ namespace Rectify11Installer
                 Wizard.SetProgressText("Restoring WinRE");
                 if (File.Exists(@"C:\Recovery\WindowsRE\Winre.wim") || File.Exists(@"C:\Windows\System32\Recovery\Winre.wim"))
                 {
-                    File.Move(@"C:\Windows\Rectify11\winre.bak.wim", @"C:\Windows\System32\Recovery\Winre.wim", true);
-                    Directory.Delete(@"C:\Recovery", true);
-                    Task.Run(() => PatcherHelper.RunAsyncCommands("reagentc.exe", "/disable", @"C:\Windows\System32"));
-                    Task.Run(() => PatcherHelper.RunAsyncCommands("reagentc.exe", "/setreimage /path " + @"C:\Windows\System32\Recovery", @"C:\Windows\System32"));
-                    Task.Run(() => PatcherHelper.RunAsyncCommands("reagentc.exe", "/enable", @"C:\Windows\System32"));
+                    if (File.Exists(@"C:\Windows\Rectify11\winre.bak.wim"))
+                    {
+                        File.Move(@"C:\Windows\Rectify11\winre.bak.wim", @"C:\Windows\System32\Recovery\Winre.wim", true);
+                        Directory.Delete(@"C:\Recovery", true);
+                        Task.Run(() => PatcherHelper.RunAsyncCommands("reagentc.exe", "/disable", @"C:\Windows\System32"));
+                        Task.Run(() => PatcherHelper.RunAsyncCommands("reagentc.exe", "/setreimage /path " + @"C:\Windows\System32\Recovery", @"C:\Windows\System32"));
+                        Task.Run(() => PatcherHelper.RunAsyncCommands("reagentc.exe", "/enable", @"C:\Windows\System32"));
+                    }
                 }
                     //Directory.Delete(@"C:\Windows\Rectify11", true);
